+++ example/lib/main.dart
@@ -1,59 +1,78 @@
-// ignore_for_file: unawaited_futures, public_member_api_docs
-import 'package:flutter/material.dart';
-import 'package:flutter_analytics/flutter_analytics.dart';
+// ignore_for_file: public_member_api_docs
 
-Future<void> example() async {
+import 'package:flutter/material.dart';
+import 'package:flutter_analytics/flutter_analytics.dart' show Analytics;
+
+void example() {
   // initial setup to run once on application lifecycle
-  Analytics.setup(destinations: ['http://localhost:3000/analytics']);
+  Analytics().setup(destinations: ['http://localhost:3000/analytics']);
 
   // uniquely identify group of users
-  Analytics.group('someGroupId', {'numTrait': 7, 'txtTrait': 'tGroup'});
+  Analytics().group('someGroupId', {'numTrait': 7, 'txtTrait': 'tGroup'});
 
   // uniquely identify current user and its traits
-  Analytics.identify('anUserId', {'numTrait': 5, 'txtTrait': 'uUser'});
+  Analytics().identify('anUserId', {'numTrait': 5, 'txtTrait': 'uUser'});
 
   // identify current screen being viewed
-  Analytics.screen('My Screen', {'numProp': -1, 'txtProp': 'pScreen'});
+  Analytics().screen('My Screen', {'numProp': -1, 'txtProp': 'pScreen'});
 
   // track discrete events
-  Analytics.track('Any Event', {'numProp': 3, 'txtProp': 'pTrack'});
+  Analytics().track('Any Event', {'numProp': 3, 'txtProp': 'pTrack'});
 
-  // debug logged events to console (returning `true` bypasses `destinations`)
-  Analytics.flush((batch) async {
+  // debug logged events to console (return `true` to bypass `destinations`)
+  Analytics().flush((batch) async {
     batch.forEach(print);
 
     return true;
   });
 }
 
-Future<String> runExample() async {
-  try {
-    await example();
+class ExampleApp extends StatefulWidget {
+  const ExampleApp();
 
-    return 'Everything fine, see console!';
-  } catch (e, s) {
-    debugPrint('$e\n$s');
+  @override
+  State<StatefulWidget> createState() => ExampleAppState();
+}
 
-    return 'Something went wrong, see console!';
+class ExampleAppState extends State<ExampleApp> with WidgetsBindingObserver {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: Scaffold(
+        body: Center(
+          child: Text(runExample()),
+        ),
+      ),
+    );
+  } // build
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) =>
+      Analytics().updateAppLifecycleState(state);
+
+  @override
+  void dispose() {
+    WidgetsBinding.instance.removeObserver(this);
+    super.dispose();
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    WidgetsBinding.instance.addObserver(this);
+  }
+
+  String runExample() {
+    try {
+      example();
+
+      return 'Everything fine, see console!';
+    } catch (e, s) {
+      debugPrint('$e\n$s');
+
+      return 'Something went wrong, see console!';
+    }
   }
 }
 
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder(
-      future: runExample(),
-      builder: (context, snapshot) {
-        return MaterialApp(
-          home: Scaffold(
-            body: Center(
-              child: Text(snapshot.data.toString() ?? 'hold on'),
-            ),
-          ),
-        );
-      }, // builder
-    );
-  } // build
-}
-
-void main() => runApp(MyApp());
+void main() => runApp(const ExampleApp());